requirements.txt
Django==2.2
django-appconf==1.0.4
django-crispy-forms==1.9.1
django-redis==4.12.1
django-select2==7.6.1
django-widget-tweaks==1.4.8
partial-date==0.1
psycopg2==2.8.6
pytz==2020.4
redis==3.5.3
sqlparse==0.4.1
